fix issues,add ts, refactor ,add tsconfig, split server and client logic
