+++ .gitignore
@@ -1,4 +1,6 @@
+lib
 # Logs
+
 logs
 *.log
 npm-debug.log*
