+++ package.json
@@ -2,12 +2,13 @@
   "name": "stackoverflowtags",
   "version": "1.0.0",
   "description": "Application add tags to stackvoverflow",
-  "main": "lib/index.js",
+  "main": "lib/server/index.js",
   "scripts": {
-    "start": "node ./src/index.js",
+    "build": "tsc -p src/server/tsconfig.json && tsc -p src/client/tsconfig.json",
+    "start": "npm run build && node ./lib/server/index.js",
     "start-debug": "cross-env DEBUG=stackoverflowtags:* npm run start",
-    "lint": "eslint ./test ./src/index.js",
-    "test": "tsc && npm run lint && mocha ./test/test.js",
+    "lint": "eslint ./test ./src/server/index.js",
+    "test": "npm run lint && mocha ./test/test.js",
     "watch": "npm test -- -w",
     "lint-fix": "npm run lint -- --fix"
   },
@@ -34,6 +35,7 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "got": "^10.6.0",
-    "qs": "^6.9.1"
+    "qs": "^6.9.1",
+    "typescript": "^3.8.3"
   }
 }
