+++ src/client/getDataFromStack.js
@@ -1,5 +1,8 @@
-async function getData () {
+import URL from '../server/url.js';
+
+export async function getData () {
     try {
+        console.log(URL + "GetData");
         const response = await fetch(URL);
         const json = await response.json();
