+++ src/client/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig",
+  "compilerOptions": {
+      "module": "system",
+      "outFile": "../../lib/client/bundle.js",
+      "allowJs": true
+  }
+}
