+++ src/server/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig",
+  "compilerOptions": {
+      "esModuleInterop": true,
+      "outDir": "../../lib/server",
+      "module": "commonjs",
+      "allowJs": true
+  }
+}
